Fix wait conditions for consent screen handling
Fix the following issues with consent screen handling:
With a very slow connection sometimes waiting for the navigation to
take place after handling the consent screen timed out.
Awaiting the navigation to take place inside the waiter method was
dirty (the method probably was meant to be for predicates which
terminate faster).
Starting to wait for the navigation after clicking the button was also
dirty, because the navigation could have taken place before calling
page.waitForNavigation().